+++ app/src/test/java/org/simple/clinic/home/HomeScreenUpdateTest.kt
@@ -5,10 +5,10 @@ import com.spotify.mobius.test.NextMatchers.hasNoEffects
 import com.spotify.mobius.test.NextMatchers.hasNoModel
 import com.spotify.mobius.test.UpdateSpec
 import com.spotify.mobius.test.UpdateSpec.assertThatNext
-import java.util.Optional
 import org.junit.Test
 import org.simple.clinic.platform.util.RuntimePermissionResult.DENIED
 import org.simple.clinic.platform.util.RuntimePermissionResult.GRANTED
+import java.util.Optional
 
 class HomeScreenUpdateTest {
 
@@ -17,7 +17,10 @@ class HomeScreenUpdateTest {
     val defaultModel = HomeScreenModel.create()
     UpdateSpec(HomeScreenUpdate())
         .given(defaultModel)
-        .whenEvent(RequestNotificationPermission(permission = Optional.of(GRANTED)))
+        .whenEvent(RequestNotificationPermission(
+            permission = Optional.of(GRANTED),
+            permissionString = "NOTIFICATION_PERMISSION"
+        ))
         .then(assertThatNext(
             hasNoModel(),
             hasNoEffects()
@@ -29,7 +32,10 @@ class HomeScreenUpdateTest {
     val defaultModel = HomeScreenModel.create()
     UpdateSpec(HomeScreenUpdate())
         .given(defaultModel)
-        .whenEvent(RequestNotificationPermission(permission = Optional.of(DENIED)))
+        .whenEvent(RequestNotificationPermission(
+            permission = Optional.of(DENIED),
+            permissionString = "NOTIFICATION_PERMISSION"
+        ))
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(ShowNotificationPermissionDenied)